fix(sketchybar): force update the logo on app switch event
I would switch apps and sketchybar would take a full second to update the bar
icon on app switch. I believe the utility runs on a cron of 1s. The problem
is that app switching is laggy to the bar. This forces the bar to read the
icon immediately on app switch.

One small problem is that the icon now switches *faster* than the bar
title does. I am not sure the right way to fix this, but I will
proabably be binding skhd to send the correct yabai commands instead of
relying on the window switch event.