fix(home-assistant): govee-local-api override with PR
The govee-local-api sometimes only half-sets the govee light state. Most
likely not the fault of the api, but still, the api should account for
the flakiness of the govee light. This PR has more aggressive strategy
for making sure the desired light state is set.

I am hopeful this PR will fix my issues!