feat(`git review`): add
This command makes it miles easier to review PRs. Combine this with `git
pr`, and reviewing PRs will be very quick. `git review` just resets the
git repo to the base branch of the PR. This means all the changed files
become working, making in-editor review extremely simple and easy.