chore(kernel, nvidia driver, and firefox): use pkgs-bleeding
In general, my opinion is that the bleeding packages for these three
will contain the least bugs. My recent firefox updates have introduced
some rendering bugs, and I think that is partly because my gpu drivers
were not being kept up-to-date with it. Therefore, I think it is best to
just update all three together.